flask_app/pyproject.toml
[tool.poetry]
name = "flask-my-blog-system"
version = "0.1.0"
description = ""
authors = ["Salmon <sos2012happy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Flask-RESTful = "^0.3.10"
Flask = "^3.0.0"
python-dotenv = "^1.0.0"
flask-restx = "^1.1.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
marshmallow-sqlalchemy = "^0.29.0"
flask-marshmallow = "^0.15.0"
flask-jwt-extended = "^4.5.3"
passlib = "^1.7.4"
flask-caching = "^2.1.0"
redis = "^5.0.1"
pyuwsgi = "^2.0.23"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
